+++ src/pages/EditUserPage/types.ts
@@ -1,7 +1,0 @@
-export interface EditUserFormData {
-	email: string;
-	password?: string;
-	confirmPassword?: string;
-	firstName?: string;
-	lastName?: string;
-}